Added inital game base
Added the code I've been working on. It includes the base game. Right now there is no lose condition. Consider creating a new variable that either tracks the time alive or the amount of pipes passed. Maybe consider a way for the game to end if a pipe is hit.